websocket/requirements.txt
python-socketio
python-engineio
aiohttp
